+++ tests/ActivityLoggerTest.php
@@ -3,13 +3,14 @@
 namespace Spatie\Activitylog\Test;
 
 use Auth;
+use Carbon\Carbon;
 use Illuminate\Support\Collection;
 use Spatie\Activitylog\Models\Activity;
 use Spatie\Activitylog\Test\Models\User;
 use Spatie\Activitylog\Test\Models\Article;
 use Spatie\Activitylog\Exceptions\CouldNotLogActivity;
 
-class ActivityloggerTest extends TestCase
+class ActivityLoggerTest extends TestCase
 {
     /** @var string */
     protected $activityDescription;
@@ -272,5 +273,28 @@ class ActivityloggerTest extends TestCase
         activity()->causedBy(null)->log('nothing');
 
         $this->markTestAsPassed();
+    }
+
+    /** @test */
+    public function it_can_log_activity_when_attributes_are_changed_with_tap()
+    {
+        $properties = [
+            'property' => [
+                'subProperty' => 'value',
+            ],
+        ];
+
+        activity()
+            ->tap(function (Activity $activity) use ($properties) {
+                $activity->properties = collect($properties);
+                $activity->created_at = Carbon::yesterday()->startOfDay();
+            })
+            ->log($this->activityDescription);
+
+        $firstActivity = Activity::first();
+
+        $this->assertInstanceOf(Collection::class, $firstActivity->properties);
+        $this->assertEquals('value', $firstActivity->getExtraProperty('property.subProperty'));
+        $this->assertEquals(Carbon::yesterday()->startOfDay()->format('Y-m-d H:i:s'), $firstActivity->created_at->format('Y-m-d H:i:s'));
     }
 }